Introduce interface KeyValueReader extends KeyValueReaderDataInputBuffer, KeyValueReaderBytesWritable.
KeyValueReaderDataInputBuffer uses DataInputBuffer.
KeyValueReaderBytesWritable uses BytesWritable.

+++ tez-common/src/main/java/org/apache/tez/common/io/NonSyncByteArrayInputStream.java
@@ -77,7 +77,7 @@ public class NonSyncByteArrayInputStream extends ByteArrayInputStream {
       k = n < 0 ? 0 : n;
     }
 
-    pos += k;
+    pos += (int)k;
     return k;
   }
